perf: Set AggressiveInlining on some functions at runtime
Turns out with some unsafe code you can just do this. It won't affect
cases where the callers have already been JITed, but will be considered
for every caller later on.

The selection of methods here is somewhat unscientific, I started with
by looking up some in a profile, and then went through the relevant
types and manually judged whether the attribute was worth adding.

The criteria here are roughly:
- Method size larger than 20b (the default inline threshold)
- Method is frequently used
- Method does not throw an exception directly (as then it couldn't be
  inlined anyway).
- Method is "simple", which is subjective